Add proper guards for platform code
Do not expect that non-ARM code is automatically X86 code as it does not
have to be. Returns failure (not thunk) in other case.

Make sure that windows test code is not called on non-windows platforms.
Test code tests/test-dladdr.c is written to be validated on any platforms,
including non-windows.