Fix is_valid_address() function
Returned structure is valid only if the function return value is same as
the size of the structure. So check for the correct size, not just for the
non-zero size.

"AllocationProtect" member contains the initial protection, valid at the
time when VirtualAlloc() was called. "Protect" member contains the current
protection of the requested address, valid at the time when VirtualQuery()
was called. So is_valid_address() function has to check "Protect" member,
not the "AllocationProtect".

If "State" member is MEM_FREE or MEM_RESERVE then "Protect" member is
undefined. So check "State" member before "Protect" member.